+++ src/app/app.module.ts
@@ -9,6 +9,7 @@ import { AutoCompleteListSampleComponent } from './autocomplete-list-sample/auto
 import { AppComponent } from './app.component';
 import { BreadcrumbSampleComponent } from './breadcrumb-sample/breadcrumb-sample.component';
 import { CheckboxSampleComponent } from './checkbox-sample/checkbox-sample.component';
+import { DatetimeSampleComponent } from './datetime-sample/datetime-sample.component';
 import { InputSampleComponent } from './input-sample/input-sample.component';
 import { FilterSampleComponent } from './filter-sample/filter-sample.component';
 import { GeolocationSampleComponent } from './geolocation-sample/geolocation-sample.component';
@@ -18,6 +19,7 @@ import { PaginationSampleComponent } from './pagination-sample/pagination-sample
 import { TooltipSampleComponent } from './tooltip-sample/tooltip-sample.component';
 import { CoreModule } from './core/core.module';
 import { PrismService } from './core/services/prism-service.service';
+import { toString } from './datetime-sample/toString.pipe';
 import { ModalSampleComponent } from './modal-sample/modal-sample.component';
 import { SelectSampleComponent } from './select-sample/select-sample.component';
 import { TooltipComponent } from './tooltip-sample/tooltip';
@@ -30,6 +32,7 @@ const appRoutes: Routes = [
   { path: 'autocomplete-list', component: AutoCompleteListSampleComponent },
   { path: 'breadcrumb', component: BreadcrumbSampleComponent },
   { path: 'checkbox', component: CheckboxSampleComponent },
+  { path: 'datetime', component: DatetimeSampleComponent },
   { path: 'index', component: MainComponent },
   { path: 'filter', component: FilterSampleComponent },
   { path: 'geolocation', component: GeolocationSampleComponent },
@@ -51,6 +54,7 @@ const appRoutes: Routes = [
     MainComponent,
     AppComponent,
     CheckboxSampleComponent,
+    DatetimeSampleComponent,
     FilterSampleComponent,
     GeolocationSampleComponent,
     MapSampleComponent,
@@ -62,7 +66,8 @@ const appRoutes: Routes = [
     BreadcrumbSampleComponent,
     ModalSampleComponent,
     RadiogroupSampleComponent,
-    VoicerecognitionSampleComponent
+    VoicerecognitionSampleComponent,
+    toString
   ],
   imports: [
     CoreModule,